preliminary results.
implemented tilts in bypass Wien-filters and looked at how they affect the invariant spin axis. the Wien-filters are modeled as RSX spin-kicks. this is of note, since AS EXPECTED, radial spin-kicks synergize with the arc vertical kicks producing longitudinal n-axis deviation (not simply vertical->radial, but vertical->radial+longitudinal).

I will now proceed turning on the RADIAL compensator. in the actual calibration procedure, two compensators are expected to be used: one radial and one longitudinal, to compensate both nx and nz. this raises theoretical concerns on my part, as to whether the EDM is supposed to be a phenomenon or a force, and if the latter, how does our so free forcing the n-axis this way and that mix with it?